backend/requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.34.0
tokenizers==0.14.1
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
SQLAlchemy==2.0.25
requests==2.32.0
